requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
bleach==2.1.2
certifi==2017.11.5
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
cryptography==2.3.1
decorator==4.2.1
defusedxml==0.5.0
dj-database-url==0.4.2
Django==2.1
django-appconf==1.0.2
django-autocomplete-light==3.3.0rc5
django-cleanup==2.1.0
django-crispy-forms==1.7.0
django-debug-toolbar==1.9.1
django-extensions==1.9.9
django-jsonfield==1.0.1
django-location-field==2.0.3
django-notify-x==0.1.9
django-tables2==1.19.0
django-widget-tweaks==1.4.1
entrypoints==0.2.3
Fabric3==1.14.post1
graphviz==0.8.2
gunicorn==19.9.0
html5lib==1.0.1
idna==2.6
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.0
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.1
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.3.1
oauthlib==2.0.6
pandocfilters==1.4.2
paramiko==2.4.1
parso==0.1.1
pendulum==1.4.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.4
ptyprocess==0.6.0
pyasn1==0.4.4
pycparser==2.18
pydotplus==2.0.2
Pygments==2.2.0
PyJWT==1.5.3
PyNaCl==1.2.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2017.3
pytzdata==2018.3
pywinpty==0.5.1
pyzmq==17.1.2
qtconsole==4.3.1
requests==2.18.4
requests-oauthlib==0.8.0
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.6.0
sqlparse==0.2.4
stripe==1.79.1
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
traitlets==4.3.2
typing==3.6.4
tzlocal==1.5.1
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.13
whitenoise==3.3.1
widgetsnbextension==3.1.0
